chore: move transformer scripts into src dir
